Show info, if update available; moved tabs to __init__.py
