+++ Rare/__init__.py
@@ -1,5 +1,5 @@
 import os
 
-__version__ = "0.9.8.3"
+__version__ = "0.9.8"
 style_path = os.path.join(os.path.dirname(__file__), "Styles/")
 lang_path = os.path.join(os.path.dirname(__file__), "languages/")
